Move layout-shell to lib/components
