+++ src/routes/(main)/households/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import LayoutShell from "../dashboard/(components)/layout-shell.svelte";
+  import LayoutShell from "$lib/components/layout-shell.svelte";
   import type { PageData } from "./$types";
   import HouseholdMember from "./household-member.svelte";
