+++ src/routes/(main)/posts/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { formatRelative } from "$lib/date";
-  import LayoutShell from "../dashboard/(components)/layout-shell.svelte";
+  import LayoutShell from "$lib/components/layout-shell.svelte";
   import type { PageData } from "./$types";
 
   export let data: PageData;
